requirements.txt
pyqt==5.15.4